rename file to a more consistent system
