+++ frontend/src/Home.js
@@ -5,17 +5,17 @@ import styled from '@emotion/styled'
 import { PhaseBanner } from './components/phase-banner'
 import { PageNotFound } from './PageNotFound'
 import { FileUploadPage } from './FileUploadPage'
-import { Suspectinfo } from './Suspectinfo'
+import { Suspectinfo } from './SuspectInfoYesNoPage'
 import { ScamInfoPage } from './ScamInfoPage'
 import { MoneyLostPage } from './MoneyLostPage'
 import { ConfirmationPage } from './ConfirmationPage'
 import { SuspectInfoPage } from './SuspectInfoPage'
 import { ContactInfoPage } from './ContactInfoPage'
-import { LoseMoney } from './LoseMoney'
+import { LoseMoney } from './MoneyLostYesNoPage'
 import { ThankYou } from './ThankYou'
-import { Willbecontacted } from './Willbecontacted'
+import { Willbecontacted } from './WillBeContactedYesNoPage'
 import { LandingPage } from './LandingPage'
-import { FilesToUpload } from './FilesToUpload'
+import { FilesToUpload } from './FileUploadYesNoPage'
 
 const Root = styled('div')`
   margin: 20pt;
