+++ frontend/src/MoneyLostYesNoPage.js
