+++ frontend/src/ThankYou.js
@@ -1,19 +1,27 @@
 /**@jsx jsx */
-import { jsx } from '@emotion/core'
+import { jsx, css } from '@emotion/core'
 import { Trans } from '@lingui/macro'
 import { H1, H2 } from './components/header'
 import { P } from './components/paragraph'
 import { A, ButtonLink } from './components/link'
 import { TrackPageViews } from './TrackPageViews'
+import { Steps } from './components/stepper'
 import { Container } from './components/container'
 import { Layout } from './components/layout'
-import { Ul } from './components/unordered-list'
-import { Li } from './components/list-item'
-import { Text } from './components/text'
+
+const topBarContainer = css`
+  display: flex;
+  width: 90%;
+  flex-direction: row;
+  margin-bottom: 20px;
+`
 
 export const ThankYou = () => (
   <Layout>
     <TrackPageViews />
+    <Container css={topBarContainer}>
+      <Steps activeStep={4} />
+    </Container>
     <H1>
       <Trans>Thank you for reporting</Trans>
     </H1>
@@ -37,38 +45,52 @@ export const ThankYou = () => (
     <H2 fontSize={[4, null, 5]}>
       <Trans>You may also want to:</Trans>
     </H2>
-    <Text>
-      <Ul>
-        <Li>
+    <P>
+      <ul>
+        <li>
           <Trans>
             Report directly to the website where the scam took place.
           </Trans>
-        </Li>
-        <Li>
+        </li>
+        <br />
+        <li>
           <Trans>
             Report to the financial institution where you sent money.
           </Trans>
-        </Li>
-        <Li>
+        </li>
+        <br />
+        <li>
           <Trans>
             Report to the credit bureaus,{' '}
             <A href="https://www.consumer.equifax.ca/personal/">Equifax</A> and{' '}
             <A href="https://www.transunion.ca/">TransUnion</A>, if your
             identity was stolen.
           </Trans>
-        </Li>
-      </Ul>
-    </Text>
+        </li>
+      </ul>
+    </P>
     <Container
-      width="305px"
-      marginTop={[1, null, 1]}
-      display="flex"
-      flexDirection="column"
-      justifyContent="space-between"
+      maxWidth="300px"
+      marginTop={[2, null, 3]}
+      css={css`
+        display: flex;
+        flex-direction: column;
+        justify-content: space-between;
+      `}
     >
-      <ButtonLink type="submit" to="/" textAlign="center">
-        <Trans> Report another scam</Trans>
-      </ButtonLink>
+      <Container
+        width="305px"
+        marginTop={[1, null, 1]}
+        css={css`
+          display: flex;
+          flex-direction: column;
+          justify-content: space-between;
+        `}
+      >
+        <ButtonLink type="submit" to="/" textAlign="center">
+          <Trans>Report another scam </Trans>
+        </ButtonLink>
+      </Container>
     </Container>
   </Layout>
 )
